Get rid of JBoss API spec artifacts to unify things with Jakarta, except...
for JAXB and JAX-RS as a lot of the optional RESTEasy artifacts depend
on them so you get them in every time you add an optional RESTEasy
artifact.

Thus, I thought it might be better to keep them for now but, carefully
exclude the others to have only one API jar.