Support custom unwrapping to allow returning only useful data
Certain endpoints even the `data` of the main result response is a “wrapper”. This allows unwrapping results with custom code.

For example User/Pass `listUsers` returns a data object with only a `keys` field. Using the new feature it can now be unwrapped using `r.data.keys` to return only the keys.